Create stdlib and modify engine
