+++ skript-stdlib/src/main/java/org/skriptlang/skript/stdlib/base/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Base classes for some of the syntax nodes in the standard library.
  */
-package org.skriptlang.skript.stdlib.syntax.base;
+package org.skriptlang.skript.stdlib.base;
