+++ skript-stdlib/src/main/java/org/skriptlang/skript/stdlib/effects/IncrementEffect.java
@@ -1,4 +1,4 @@
-package org.skriptlang.skript.stdlib.syntax.effects;
+package org.skriptlang.skript.stdlib.effects;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -13,24 +13,24 @@ import org.skriptlang.skript.api.util.TypeOperationUtils;
 
 import java.util.List;
 
-public class DecrementEffect implements EffectNode {
-	public static final EffectNodeType<DecrementEffect> TYPE = new EffectNodeType<>() {
+public class IncrementEffect implements EffectNode {
+	public static final EffectNodeType<IncrementEffect> TYPE = new EffectNodeType<>() {
 		@Override
 		public List<String> getSyntaxes() {
-			return List.of("decrement <expr> [by <expr>]");
+			return List.of("increment <expr> [by <expr>]");
 		}
 
 		@Override
-		public @NotNull DecrementEffect create(List<SyntaxNode> children, int matchedPattern) {
+		public @NotNull IncrementEffect create(List<SyntaxNode> children, int matchedPattern) {
 			ExpressionNode<?> amountSelector = children.size() == 2 ? (ExpressionNode<?>) children.get(1) : null;
-			return new DecrementEffect((ExpressionNode<?>) children.getFirst(), amountSelector);
+			return new IncrementEffect((ExpressionNode<?>) children.getFirst(), amountSelector);
 		}
 	};
 
 	private final @NotNull ExpressionNode<?> receiverSelector;
 	private final @Nullable ExpressionNode<?> amountSelector;
 
-	public DecrementEffect(@NotNull ExpressionNode<?> receiverSelector, @Nullable ExpressionNode<?> amountSelector) {
+	public IncrementEffect(@NotNull ExpressionNode<?> receiverSelector, @Nullable ExpressionNode<?> amountSelector) {
 		this.receiverSelector = receiverSelector;
 		this.amountSelector = amountSelector;
 	}
@@ -38,12 +38,9 @@ public class DecrementEffect implements EffectNode {
 	@Override
 	public @NotNull ExecuteResult execute(@NotNull ExecuteContext context) {
 
-		return TypeOperationUtils.applyDecrement(
-			receiverSelector.resolve(context),
-			amountSelector == null ? null : amountSelector.resolve(context).toValue()
-		)
+		return TypeOperationUtils.applyIncrement(receiverSelector.resolve(context), amountSelector == null ? null : amountSelector.resolve(context).toValue())
 			? ExecuteResult.success()
-			: ExecuteResult.failure(new ErrorValue("Failed to decrement value - does it support decrementing?"));
+			: ExecuteResult.failure(new ErrorValue("Failed to increment value - does it support incrementing?"));
 
 	}
 }
